fix: move use-input to root-layer
