+++ components/use-input/index.ts
@@ -1,0 +1,4 @@
+import useInput from './use-input'
+
+export type { BindingsChangeTarget } from './use-input'
+export default useInput
